feat(online-banking-PL): add terms and conditions
