+++ packages/lib/src/components/OnlinebankingPL/index.ts
@@ -1,7 +1,1 @@
-import IssuerListContainer from '../helpers/IssuerListContainer';
-
-class OnlineBankingPL extends IssuerListContainer {
-    public static type = 'onlineBanking_PL';
-}
-
-export default OnlineBankingPL;
+export { default } from './OnlineBankingPL';
