+++ packages/lib/src/language/locales/en-US.json
@@ -243,6 +243,7 @@
     "upi.vpaWaitingMessage": "Open your UPI app to confirm the payment",
     "upi.modeSelection": "Make a selection on how you would like to use UPI.",
     "onlineBanking.termsAndConditions": "By continuing you agree with the %#terms and conditions%#",
+    "onlineBankingPL.termsAndConditions": "By continuing you agree with the %#regulations%# and %#information obligation%# of Przelewy24",
     "ctp.loading.poweredByCtp": "Powered by Click to Pay",
     "ctp.loading.intro": "We are checking to see if you have any saved cards with Click to Pay...",
     "ctp.login.title": "Continue to Click to Pay",
