+++ packages/lib/src/language/locales/pl-PL.json
@@ -237,6 +237,7 @@
     "upi.vpaWaitingMessage": "Otwórz aplikację UPI, aby potwierdzić płatność",
     "upi.modeSelection": "Dokonaj wyboru sposobu korzystania z UPI.",
     "onlineBanking.termsAndConditions": "Kontynuując, zgadzasz się z %#Warunkami świadczenia usług%#.",
+    "onlineBankingPL.termsAndConditions": "Kontynuując, zgadzasz się z %#regulaminem%# i %#obowiązkiem informacyjnym%# serwisu Przelewy24",
     "ctp.loading.poweredByCtp": "Działa w oparciu o usługę Click to Pay",
     "ctp.loading.intro": "Sprawdzamy, czy masz zapisane karty Click to Pay...",
     "ctp.login.title": "Przejdź do Click to Pay",
